+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/hardware/HardwarexTest.kt
@@ -20,12 +20,10 @@ import android.Manifest
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import me.panpf.androidx.Androidx
-import me.panpf.androidx.app.Permissionx
-import me.panpf.androidx.hardware.Hardwarex
 import me.panpf.androidxkt.app.isGrantPermissions
 import me.panpf.androidxkt.hardware.*
-import me.panpf.javax.lang.Stringx
-import me.panpf.javax.util.Comparisonx
+import me.panpf.javaxkt.lang.isSafe
+import me.panpf.javaxkt.net.isMacAddress
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -34,48 +32,28 @@ import org.junit.runner.RunWith
 class HardwarexTest {
 
     @Test
-    fun testDeviceModel() {
-        Assert.assertTrue(getDeviceModel().isNotEmpty())
-    }
-
-    @Test
-    fun testDeviceName() {
-        Assert.assertTrue(getDeviceName().isNotEmpty())
-    }
-
-    @Test
-    fun testHardware() {
-        Assert.assertTrue(getHardware().isNotEmpty())
-    }
-
-    @Test
-    fun testSupportedAbis() {
-        Assert.assertTrue(getSupportedAbis().isNotEmpty())
-    }
-
-    @Test
     fun testDeviceId() {
         val context = InstrumentationRegistry.getContext()
         if (Androidx.isAtLeast6_0()) {
-            if (Permissionx.isGrantPermissions(context, Manifest.permission.READ_PHONE_STATE)) {
+            if (context.isGrantPermissions(Manifest.permission.READ_PHONE_STATE)) {
                 val deviceId = context.getDeviceId()
-                Assert.assertTrue("deviceId: $deviceId", Stringx.isSafe(deviceId)
-                        && !Comparisonx.areEqual(deviceId, "unknown")
-                        && !Comparisonx.areEqual(deviceId, "PermissionDenied"))
+                Assert.assertTrue("deviceId: $deviceId", deviceId.isSafe()
+                        && deviceId != "unknown"
+                        && deviceId != "PermissionDenied")
             } else {
                 val deviceId = context.getDeviceId()
                 Assert.assertEquals("deviceId: $deviceId", deviceId, "PermissionDenied")
             }
         } else {
             val deviceId = context.getDeviceId()
-            Assert.assertTrue("deviceId: $deviceId", Stringx.isSafe(deviceId))
+            Assert.assertTrue("deviceId: $deviceId", deviceId.isSafe())
         }
     }
 
     @Test
     fun testAndroidId() {
-        val androidId = Hardwarex.getAndroidId(InstrumentationRegistry.getContext())
-        Assert.assertTrue("androidId: $androidId", Stringx.isSafe(androidId))
+        val androidId = InstrumentationRegistry.getContext().getAndroidId()
+        Assert.assertTrue("androidId: $androidId", androidId.isSafe())
     }
 
     @Test
@@ -84,16 +62,16 @@ class HardwarexTest {
         if (Androidx.isAtLeast6_0()) {
             if (context.isGrantPermissions(Manifest.permission.READ_PHONE_STATE)) {
                 val subscriberId = context.getSubscriberId()
-                Assert.assertTrue("subscriberId: $subscriberId", Stringx.isSafe(subscriberId)
-                        && !Comparisonx.areEqual(subscriberId, "unknown")
-                        && !Comparisonx.areEqual(subscriberId, "PermissionDenied"))
+                Assert.assertTrue("subscriberId: $subscriberId", subscriberId.isSafe()
+                        && subscriberId != "unknown"
+                        && subscriberId != "PermissionDenied")
             } else {
                 val subscriberId = context.getSubscriberId()
                 Assert.assertEquals("subscriberId: $subscriberId", subscriberId, "PermissionDenied")
             }
         } else {
             val subscriberId = context.getSubscriberId()
-            Assert.assertTrue("subscriberId: $subscriberId", Stringx.isSafe(subscriberId))
+            Assert.assertTrue("subscriberId: $subscriberId", subscriberId.isSafe())
         }
     }
 
@@ -103,35 +81,16 @@ class HardwarexTest {
         if (Androidx.isAtLeast6_0()) {
             if (context.isGrantPermissions(Manifest.permission.READ_PHONE_STATE)) {
                 val simSerialNumber = context.getSimSerialNumber()
-                Assert.assertTrue("simSerialNumber: $simSerialNumber", Stringx.isSafe(simSerialNumber)
-                        && !Comparisonx.areEqual(simSerialNumber, "unknown")
-                        && !Comparisonx.areEqual(simSerialNumber, "PermissionDenied"))
+                Assert.assertTrue("simSerialNumber: $simSerialNumber", simSerialNumber.isSafe()
+                        && simSerialNumber != "unknown"
+                        && simSerialNumber != "PermissionDenied")
             } else {
                 val simSerialNumber = context.getSimSerialNumber()
                 Assert.assertEquals("simSerialNumber: $simSerialNumber", simSerialNumber, "PermissionDenied")
             }
         } else {
             val simSerialNumber = context.getSimSerialNumber()
-            Assert.assertTrue("simSerialNumber: $simSerialNumber", Stringx.isSafe(simSerialNumber))
-        }
-    }
-
-    @Test
-    fun testSerial() {
-        val context = InstrumentationRegistry.getContext()
-        if (Androidx.isAtLeastP()) {
-            if (context.isGrantPermissions(Manifest.permission.READ_PHONE_STATE)) {
-                val serial = Hardwarex.getSerial()
-                Assert.assertTrue("serial: $serial", Stringx.isSafe(serial)
-                        && !Comparisonx.areEqual(serial, "unknown")
-                        && !Comparisonx.areEqual(serial, "PermissionDenied"))
-            } else {
-                val serial = Hardwarex.getSerial()
-                Assert.assertEquals("serial: $serial", serial, "PermissionDenied")
-            }
-        } else {
-            val serial = Hardwarex.getSerial()
-            Assert.assertTrue("serial: $serial", Stringx.isSafe(serial))
+            Assert.assertTrue("simSerialNumber: $simSerialNumber", simSerialNumber.isSafe())
         }
     }
 
@@ -141,16 +100,16 @@ class HardwarexTest {
         if (Androidx.isAtLeast6_0()) {
             if (context.isGrantPermissions(Manifest.permission.ACCESS_WIFI_STATE)) {
                 val macAddress = context.getMacAddress()
-                Assert.assertTrue("macAddress: $macAddress", Stringx.isSafe(macAddress)
-                        && !Comparisonx.areEqual(macAddress, "unknown")
-                        && !Comparisonx.areEqual(macAddress, "PermissionDenied"))
+                Assert.assertTrue("macAddress: $macAddress", macAddress.isSafe()
+                        && macAddress != "unknown"
+                        && macAddress != "PermissionDenied")
             } else {
                 val macAddress = context.getMacAddress()
                 Assert.assertEquals("macAddress: $macAddress", macAddress, "PermissionDenied")
             }
         } else {
             val macAddress = context.getMacAddress()
-            Assert.assertTrue("macAddress: $macAddress", Stringx.isSafe(macAddress))
+            Assert.assertTrue("macAddress: $macAddress", macAddress.isMacAddress())
         }
     }
 }